Shift code to puzzle api controller
